Add Moderate difficulty problem. 4.5
Annoying thing about this code was that you'd expect the period about which the string returns to it's base form to be along the lines of "string.length()-1" or something. 
However, there's a notable case where string.length() returns a number of the form 2+3*n, where n is a positive integer. I eventually found a solution to these, but then realized that there were other, non-obvious exceptions.
In my rage, I have brute forced finding the period. Yes, I am ashamed.